backend/requirements.txt
Django>=5.2
djangorestframework
Pillow
psycopg2-binary
django-cors-headers
gunicorn
whitenoise
dj-database-url
